test: musl を対象にしたビルドでもテストがコンパイルできるようにする
TIOCSTI の定数の型が libc の実装で違う。glibc では c_ulong、musl では
c_int なので、musl を対象にすると (1 << 32) が定数の型のまま評価されて
桁あふれでコンパイルが止まっていた。

シフトを 64 ビットで書いて型を揃える。glibc 側では同じ変換が「不要な
キャスト」として clippy に拒まれるため、その 1 行だけ lint を外す。
キャストを消すと musl が通らなくなるので、外すのは lint の側になる。

+++ tests/launch.rs
@@ -887,7 +887,12 @@ fn tiocsti_is_denied_with_eperm() {
 #[test]
 fn tiocsti_with_high_bits_is_denied_with_eperm() {
     let (home, workspace) = home_with_workspace();
-    let request = libc::TIOCSTI | (1 << 32);
+    // `TIOCSTI` is a `c_ulong` against glibc and a `c_int` against musl, so the shift is
+    // written in 64 bits rather than in the constant's own type. The cast is what makes
+    // the musl build compile; against glibc the same cast is the one clippy calls
+    // unnecessary, so the lint is turned off here rather than the cast removed.
+    #[allow(clippy::unnecessary_cast)]
+    let request = libc::TIOCSTI as u64 | (1u64 << 32);
 
     let output = run_script(
         &home,
